Refactorings based on code review
