+++ .bettercodehub.yml
@@ -1,0 +1,1 @@
+component_depth: 2
